refactor: COUNT 쿼리 반환 타입 int → long 으로 변경

+++ src/main/java/com/groute/groute_server/report/adapter/in/web/dto/ReportGaugeResponse.java
@@ -7,7 +7,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 /** 리포트 게이지 조회 응답 DTO. {@link ReportGaugeView}의 Web 표현. */
 @Schema(description = "리포트 게이지 조회 응답")
 public record ReportGaugeResponse(
-        @Schema(description = "마지막 리포트 생성 이후 완료된 심화기록 수", example = "7") int currentCount,
+        @Schema(description = "마지막 리포트 생성 이후 완료된 심화기록 수", example = "7") long currentCount,
         @Schema(description = "다음 생성 기준 수", example = "10") int nextThreshold,
         @Schema(description = "currentCount / nextThreshold. 1.0 초과 가능", example = "0.7")
                 double progressRate,
